+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## 1.3.8 (2024-03-xx)
 ### Common
 * **core**: Use Google Map API for getting timezone and country from latitude and longitude
+* **core**: Endpoint `GET /classifier-job/:id/summary` receive params `limit, offset, sort, order, keyword`
+* **core**: New endpoint `GET /classifier-job/:id/validation` for getting validation status
+
+### Features
+* **core**: GET `classifier-jobs/:id` now includes totalDistinctClassifications field with the number of linked classifications
 
 ## 1.3.7 (2024-02-xx)
 ### Common